Disk should returns the size of the disk
Disk key should give the size of the first disk in GB, not the number of hard
drive. During the nova boot, nova will compare this value with the
minimal mandatory size associated with the Glance image.

Without this patch, nova will always fail and returns the "no valid host was found"
message as soon as the size is greater than 1.